complete the transformation to the new way of doing sharing and mounting
The process now goes:
 processor-> env-> mount
 util-> provider-> mount (this is where the logic to go netfs exists)
 processor-> env-> share-> add
 util-> provider-> share
