+++ commands/stop.go
@@ -3,6 +3,7 @@ package commands
 import (
 	"github.com/spf13/cobra"
 
+	"github.com/nanobox-io/nanobox/commands/registry"
 	"github.com/nanobox-io/nanobox/processors"
 	"github.com/nanobox-io/nanobox/util/display"
 )
@@ -23,5 +24,6 @@ dev and sim platforms.
 
 // stopFn ...
 func stopFn(ccmd *cobra.Command, args []string) {
+	registry.Set("keep-share", true)
 	display.CommandErr(processors.Stop())
 }
