+++ util/provider/share/share_darwin.go
@@ -2,10 +2,13 @@ package share
 
 import (
 	"fmt"
+	"flag"
 	"bytes"
 	"io/ioutil"
+	"os/exec"
 	"strings"
 
+	"github.com/jcelliott/lumber"
 
 	"github.com/nanobox-io/nanobox/models"
 )
@@ -56,7 +59,10 @@ func Add(path string) error {
 	}
 	
 	// save
-	return ioutil.WriteFile(EXPORTSFILE, []byte(strings.Join(lines, "\n")), 0644)
+	if err := ioutil.WriteFile(EXPORTSFILE, []byte(strings.Join(lines, "\n")), 0644); err != nil {
+		return err
+	}
+	return reloadServer()
 }
 
 func Remove(path string) error {
@@ -93,5 +99,36 @@ func Remove(path string) error {
 	}
 	
 	// save
-	return ioutil.WriteFile(EXPORTSFILE, []byte(strings.Join(newLines, "\n")), 0644)
+	if err := ioutil.WriteFile(EXPORTSFILE, []byte(strings.Join(newLines, "\n")), 0644); err != nil {
+		return err
+	}
+	
+	return reloadServer()
+}
+
+// reloadServer will reload the nfs server with the new export configuration
+func reloadServer() error {
+
+	// dont reload the server when testing
+	if flag.Lookup("test.v") != nil {
+      return nil
+  }
+	// TODO: make sure nfsd is enabled
+
+	// check the exports to make sure a reload will be successful; TODO: provide a
+	// clear message for a direction to fix
+	cmd := exec.Command("nfsd", "checkexports")
+	if b, err := cmd.CombinedOutput(); err != nil {
+	       lumber.Debug("checkexports: %s", b)
+	       return fmt.Errorf("checkexports: %s %s", b, err.Error())
+	}
+
+	// update exports; TODO: provide a clear error message for a direction to fix
+	cmd = exec.Command("nfsd", "update")
+	if b, err := cmd.CombinedOutput(); err != nil {
+	       lumber.Debug("update: %s", b)
+	       return fmt.Errorf("update: %s %s", b, err.Error())
+	}
+
+	return nil
 }
